tel/requirements.txt
httpx==0.23.0
python-telegram-bot==20.0a1
pydantic

databases
psycopg2-binary
asyncpg
sqlalchemy